SLIME 3.1

+++ slime/include/SimpSolver.h
@@ -115,7 +115,7 @@ class SimpSolver : public Solver {
 
         // TODO: are 64-bit operations here noticably bad on 32-bit platforms? Could use a saturating
         // 32-bit implementation instead then, but this will have to do for now.
-        uint64_t cost(Var x) const { return (uint64_t)n_occ[toInt(mkLit(x))] * (uint64_t)n_occ[toInt(~mkLit(x))]; }
+        long cost(Var x) const { return (long)n_occ[toInt(mkLit(x))] * (long)n_occ[toInt(~mkLit(x))]; }
         bool operator()(Var x, Var y) const { return cost(x) < cost(y); }
 
         // TODO: investigate this order alternative more.
@@ -135,7 +135,7 @@ class SimpSolver : public Solver {
     //
     long elimorder;
     bool use_simplification;
-    vec<uint32_t> elimclauses;
+    vec<long> elimclauses;
     vec<char> touched;
     OccLists<Var, vec<CRef>, ClauseDeleted> occurs;
     vec<long> n_occ;
